requirements.txt
psycopg2-binary
Adafruit-DHT
notifymail
